# Improve performance of Text.compare_to