--- a/engine/runtime/src/main/java/org/enso/interpreter/runtime/data/text/Text.java
+++ b/engine/runtime/src/main/java/org/enso/interpreter/runtime/data/text/Text.java
@@ -1,6 +1,7 @@
 package org.enso.interpreter.runtime.data.text;
 
 import com.ibm.icu.text.BreakIterator;
+import com.ibm.icu.text.Normalizer2;
 import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.dsl.Cached;
 import com.oracle.truffle.api.interop.InteropLibrary;
@@ -26,6 +27,13 @@ public final class Text implements TruffleObject {
   private static final Lock LOCK = new ReentrantLock();
   private volatile Object contents;
   private volatile int length = -1;
+  private volatile FcdNormalized fcdNormalized = FcdNormalized.UNKNOWN;
+
+  private enum FcdNormalized {
+    YES,
+    NO,
+    UNKNOWN
+  }
 
   private Text(String string) {
     this.contents = string;
@@ -55,6 +63,32 @@ public final class Text implements TruffleObject {
       length = l;
     }
     return l;
+  }
+
+  @Builtin.Method(description = """
+  Checks whether this text is in FCD normalized form.
+
+  > Example
+    Check if the string is normalized
+
+        "14.95€".is_normalized
+  """)
+  public boolean is_normalized() {
+    switch (fcdNormalized) {
+      case YES -> {
+        return true;
+      }
+      case NO -> {
+        return false;
+      }
+      case UNKNOWN -> {
+        Normalizer2 normalizer = Normalizer2.getInstance(null, "nfc", Normalizer2.Mode.FCD);
+        boolean isNormalized = normalizer.isNormalized(toString());
+        setFcdNormalized(isNormalized);
+        return isNormalized;
+      }
+    }
+    return false;
   }
 
   /**
@@ -187,6 +221,14 @@ public final class Text implements TruffleObject {
     this.contents = contents;
   }
 
+  private void setFcdNormalized(boolean flag) {
+    if (flag) {
+      fcdNormalized = FcdNormalized.YES;
+    } else {
+      fcdNormalized = FcdNormalized.NO;
+    }
+  }
+
   @Override
   public String toString() {
     Object c = this.contents;
@@ -243,4 +285,5 @@ public final class Text implements TruffleObject {
     }
     return result;
   }
+
 }